api/requirements.txt
fastapi
uvicorn
pydantic>=2.0.0
sqlmodel>=0.0.22
sqlalchemy
langgraph
langchain-core
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
httpx
psycopg2-binary
bcrypt
google-generativeai
